Addressed review comments.
Let's address the review comments before we merge the request. - Replaced all tabs with white spaces
Need to understand the purpose of this Abstract class? - Removed abstract class
Need to update the name "Architectural layer should not be skipped". May be something which talks about public getter. - Corrected the message
Should this method be renamed to isPublicGetterForRestrictedType- Dobe
Does this method give false positive for non getter public method? Should we check for "get" prefix in the method? Also may be we need to check return type of the method to check if return type is a restricted type. - Done
Should this be debug logging? seem like we are checking for debug level. - Done
There is no Annotation like @facade - Removed @Facase annotation
Please make this method public and run the test. - Done